+++ sapling_generators.py
@@ -5,7 +5,7 @@ from pyblake2 import blake2s
 
 from sapling_jubjub import Point, JUBJUB_COFACTOR
 from tv_output import render_args, render_tv
-from sapling_utils import i2leosp
+from utils import i2leosp
 
 # First 64 bytes of the BLAKE2s input during group hash.
 # This is chosen to be some random string that we couldn't have
