+++ sapling_merkle_tree.py
@@ -4,7 +4,7 @@ import sys; assert sys.version_info[0] >= 3, "Python 3 required."
 from binascii import unhexlify
 
 from sapling_pedersen import pedersen_hash
-from sapling_utils import i2lebsp, leos2bsp
+from utils import i2lebsp, leos2bsp
 
 MERKLE_DEPTH = 32
