+++ sapling_notes.py
@@ -7,7 +7,7 @@ from sapling_pedersen import (
     mixing_pedersen_hash,
     windowed_pedersen_commitment,
 )
-from sapling_utils import i2lebsp
+from utils import i2lebsp
 
 def note_commit(rcm, g_d, pk_d, v):
     return windowed_pedersen_commitment(rcm, [1] * 6 + i2lebsp(64, v) + g_d + pk_d)
