+++ sapling_zip32.py
@@ -5,7 +5,7 @@ from pyblake2 import blake2b
 
 from sapling_key_components import to_scalar, prf_expand, diversify_hash, DerivedAkNk, DerivedIvk
 from sapling_generators import SPENDING_KEY_BASE, PROVING_KEY_BASE
-from sapling_utils import i2leosp, i2lebsp, lebs2osp
+from utils import i2leosp, i2lebsp, lebs2osp
 from ff1 import ff1_aes256_encrypt
 from tv_output import render_args, render_tv, option, Some
